+++ template/src/index.js
@@ -5,6 +5,7 @@ const path = require("path");
 const os = require("os");
 const ErrorHandler = require("./error-handler");
 const ProcessManager = require("./process-manager"); // <-- import ProcessManager
+const ServerUtils = require("./server-utils"); // 추가
 
 // Application State Management with better error handling
 class AppState {
@@ -88,112 +89,6 @@ class AppState {
 // Initialize global state
 const appState = new AppState();
 
-// Server utilities with optimized retry logic
-class ServerUtils {
-  static async checkServerStatus(url, timeout = 3000) {
-    return ErrorHandler.handleAsyncError(
-      "ServerUtils.checkServerStatus",
-      async () => {
-        const controller = new AbortController();
-        const timeoutId = setTimeout(() => controller.abort(), timeout);
-
-        try {
-          const res = await fetch(url, {
-            method: "HEAD",
-            signal: controller.signal,
-          });
-          clearTimeout(timeoutId);
-          return res.status === 200;
-        } catch (error) {
-          clearTimeout(timeoutId);
-          throw error;
-        }
-      },
-      false
-    );
-  }
-
-  static async waitFor(milliseconds) {
-    return new Promise((resolve) => {
-      setTimeout(resolve, milliseconds);
-    });
-  }
-
-  // Optimized retry logic with exponential backoff
-  static async waitForServerWithExponentialBackoff(
-    url,
-    maxAttempts = 10,
-    initialDelay = 1000,
-    maxDelay = 5000,
-    backoffFactor = 1.5
-  ) {
-    let currentDelay = initialDelay;
-
-    for (let attempt = 1; attempt <= maxAttempts; attempt++) {
-      // First attempt happens immediately
-      if (attempt > 1) {
-        await this.waitFor(currentDelay);
-      }
-
-      const isUp = await this.checkServerStatus(url);
-      if (isUp) {
-        console.log(
-          `Server responded on attempt ${attempt} after ${currentDelay}ms delay`
-        );
-        return true;
-      }
-
-      // Calculate next delay with exponential backoff
-      currentDelay = Math.min(currentDelay * backoffFactor, maxDelay);
-
-      console.log(
-        `Server check attempt ${attempt}/${maxAttempts} failed. Next delay: ${currentDelay}ms`
-      );
-    }
-
-    return false;
-  }
-
-  // Alternative: Smart retry with jitter to prevent thundering herd
-  static async waitForServerWithJitter(
-    url,
-    maxAttempts = 10,
-    baseDelay = 100,
-    maxDelay = 3000
-  ) {
-    for (let attempt = 1; attempt <= maxAttempts; attempt++) {
-      if (attempt > 1) {
-        // Add randomness to prevent multiple instances hitting server simultaneously
-        const jitter = Math.random() * 0.3; // 30% jitter
-        const delay = Math.min(
-          baseDelay * Math.pow(2, attempt - 2) * (1 + jitter),
-          maxDelay
-        );
-        await this.waitFor(delay);
-      }
-
-      const isUp = await this.checkServerStatus(url);
-      if (isUp) {
-        return true;
-      }
-    }
-
-    return false;
-  }
-
-  // Enhanced server startup with multiple strategies
-  static async waitForServerSmart(url, strategy = "exponential") {
-    switch (strategy) {
-      case "exponential":
-        return await this.waitForServerWithExponentialBackoff(url);
-      case "jitter":
-        return await this.waitForServerWithJitter(url);
-      default:
-        return await this.waitForServerWithExponentialBackoff(url);
-    }
-  }
-}
-
 // 1. Start R process
 async function startShinyProcessWithState(appState) {
   try {
@@ -213,7 +108,7 @@ async function waitForServerReady(appState, strategy = "exponential") {
 }
 
 // 3. Calculate retry delay
-function getRetryDelay(attempt, base = 1000, factor = 1.5, max = 10000) {
+function getRetryDelay(attempt, base = 2000, factor = 1.5, max = 10000) {
   return Math.min(base * Math.pow(factor, attempt), max);
 }
